refactor: rename process component to 'process sumary' in order to use this name to the process detail component
