+++ src/components/ProcessRoster/ProcessRoster.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import Process from '../Process/Process';
+import ProcessSummary from '../ProcessSummary/ProcessSummary';
 
 import { List, ListItem } from './ProcessRoster.styles';
 
@@ -12,7 +12,7 @@ const ProcessRoster = ({ processList }) => {
       {processList.map(process => {
         return (
           <ListItem key={process.id}>
-            <Process process={process} />
+            <ProcessSummary process={process} />
           </ListItem>
         );
       })}
